Extracted file signal class from HBaseSourceITCase
